NDD-000: Address node installation issues.
`Shrinkwrap` the npm dependencies for continuous integration and local development. Some changes happened in the past week that are causing issues and this should give us a consistent set-up.

- Test `npm install` and `gulp` after wiping the `node_modules` folder in the theme or starting from scratch.

Prior to this, the `node-sass` post install script would fail. Even if that proceeded, the `gulp` command would fail because of a `SyntaxError: Use of const in strict mode.` error.

See https://docs.npmjs.com/files/shrinkwrap.json